Script to generate a PostgreSQL schema from GTFS
The script acts on a directory with the extracted contents of a GTFS
file. It checks which files are present and prepares one table for each
of them. Then it checks which columns are present and in which order and
generates a column defintiion for each of them. Finally, the contents of
the CSV files are imported into the database using PostgreSQL's \COPY
mode.